Create Jaguar2 JUnit Example module
The example is based on the program `Max` published in "An efficient
bitwise algorithm for intra-procedural data-flow testing coverage",
Information Processing Letters. Volume 113, Issue 8, April, 2013. Pages
293-300.

The tests are based on the test suite from program `Max` published in
"Evaluating data-flow coverage in spectrum-based fault localization", in
2019 ACM/IEEE International Symposium on Empirical Software Engineering
and Measurement, ESEM 2019, Porto de Galinhas, Recife, Brazil, September
19-20, 2019. Pages 1–11.

We set Maven Surefire Plugin to ignore failures during testing on this
module. Note that some tests are supposed to fail, but we don't want
that the build exits with an error code.