Jaguar2 JaCoCo Provider `analyze` implementation
This commit is the initial implementation for Jaguar2 JaCoCo Provider
analysis and includes custom Spectrum interfaces implementations for use
with JaCoCo and an initial analysis implementation.

We also extracted an abstract `CoverageController` class called
`ClassFilesController` to Jaguar2 Commons module, an abstract controller
used for analysis that requires the class files (probably required for
Jaguar2 BA-DUA Provider).