Removes `skip` config of Apache Maven Install Plugin for Jaguar2 Example JUnit JaCoCo
Same of 30a02a3 but for `jaguar2-example-junit-jacoco` module.

We was skipping it because this module was not generating any artifact
and without this configuration the build was failing.

See: PR #33

But now this module has source files and is generating artifacts, So we
can revert this configuration (reverts e2028f0).