Do not create test jar for Jaguar2 Example JUnit
The test jar was introduced on PR #11 (cbb2561) to avoid code
duplications, so we re-use the tests on other test modules.

We are not using the test jar since PR #40 (696923f, a61c003) in order
to use symblinks. So now we don't need to generate the test jar anymore.

Actually, this commit removes the entire Apache Maven JAR Plugin
configuration because we use it only to generate the test jar.