Jaguar2 Example JUnit BA-DUA do not inherit Jaguar2 Parent POM
This is basically the same of PR #42 (75ff813) but for Jaguar2 Example
JUnit BA-DUA module.

Note that we removed the `license.header.fileLocation` property because
this property now is useless.

See: https://github.com/saeg/jaguar2/pull/42#issuecomment-1600943264

For the same reason we removed the JaCoCo Maven Plug-in configuration,
and also because we don't wanna this module to have JaCoCo agent
running. Note that the plugin was configured to skip its execution.