Removes duplications and noisy configs from Jaguar2 Examples POMs
Removes duplications and noisy configs from Jaguar2 Examples POMs that
was introduced by PR #42, #43 and #44.

We basically did all this refactoring (PR #36, #40, #42, #43 and #44) to
make the Jaguar2 Examples Maven POMs more realistic (more similar with
real world usage scenarios).

But these changes introduced a lot of noisy code (profiles and other
build configurations), and wrost, code duplications on POM
configurations. All required to build the examples modules with all
supported JDKs.

I know that if examples modules do not inheriting any configuration from
parent POM they will be the more "realistic". I mean, simply and easy to
reproduce, because they won't rely on implicit/inherited configurations.
On the other hand, we want that the examples POMs to be more easy to
read/follow as possible.

To achieve this, the examples POMs now will inherit a parent POM
(Jagaur2 Examples Parent Module) with only the duplicated/noisy
configurations (just to get rid of it on the actual examples modules) so
the examples modules will be as simple as possible.

As you can see, these changes removes the property
`project.build.sourceEncoding` that configures UTF-8 character encoding
scheme, not really important for the examples. Also removes the default
properties (and profiles) for `maven.compiler.source` and
`maven.compiler.target` that are used for builds with different/newer
JDKs (pretty straightforward to configure and not relevant for the
example).