Moved example dependencies versions to parent POM
Following the same principle of the previous commit (ce506a3). Now we
are moving all dependencies versions from examples to dependency
management on the parent POM.

Again, these versions are not really important for the examples. For
now, the only requirement for Jaguar2 is JUnit 4, but minor and patch
versions are not relevant for the examples, so we can remove dependency
versions entirely.