Moved example plugins versions to parent POM
Again, following the same principle of the previous commits (ce506a3,
0f1d721). Now I'm are moving all plugins versions from examples to plugin
management on the parent POM.

And again, these versions are not really important for the examples.
Probably Jaguar2 will not work with very old versions of Maven Surefire
Plugin or JaCoCo Maven plug-in due to API incompatibilities. But I'm not
worried about incompatibilities with very old version, but more
concerned about promoting usage of an old version because we didn't
update the version on the example POMs.