Move `testFailureIgnore` on examples to parent POM
One more time, following the same principle of the previous commits
(ce506a3, 0f1d721, 205d98d, 6f16ca0). Now moved to the exmaples parent
POM the configuration that makes Maven Surefire Plugin to ignore
failures during tests. The examples tests are failing on purpose (see
commit 31811b8 of PR #8) and we want to ignore these failures to not
stop our build. For us this is not a problem since we have test
validations on this (see PR #34). On the other hand, users of Jaguar2
will probably not want to ignore test failures, so good idea to somehow
"hide" this configuration from the examples.