Sort spectrum traversal to ensure build is reproducible for different Java versions
The build of the previous commit (da1dac4) failed on Java 6 and 7
because the contents of the XML and CSV reports are not on same order
the validation expect. Probably because there is some differences on the
hashmap implementation on these Java versions that affects iteration
order when compared with newer Java versions.

This was not a problem when the validation had only a single class. But
this changed on commit cea588f.

Since we can't rely on the hashmap traversal order and we can't simply
use a linked hashmap because there is no guarantee that we will put data
in any specific order, we are sorting the data before the treaversal.

Note that we can't guarantee that we will put data in any specific order
because we are feeding the data based on the list of files in a
directory and this not guaranteed to appear in any specifc order.