Removed interface `ISpectrumVisitor`
This will simplify the integration of Jaguar with spectrum exporters.

The `CoverageController` implementations will now return a collection of
`IClassSpectrum` directly to the caller.

Also fixed the JavaDoc for `Jaguar#testRunFinished`. The throws
`IOException` was introduced by PR #29 (33504af) but changed to
`Exception` on PR #30 (2713675) and the JavaDoc wasn't updated
accordingly.