Create interface `ISourceFileSpectrum`
This commit also create the implementations of this interfaces for
Jaguar2 JaCoCo Provider. The `BundleSpectrum` is responsible for group
all `SourceFileSpectrum` within a respective `PackageSpectrum`. Note
that now the interface `IPackageSpectrum` provides a new method to
access the list of `ISourceFileSpectrum`within the respective package
along with the other method that provides access to the list of
`IClassSpectrum` within the same package.

The `SpectrumBuilder` from Jaguar2 JaCoCo Provider is the one that
instantiate implementations of this interface based on the source file
name from the respective `ClassSpectrum`. That's why we added a new
method on `IClassSpectrum` to get the respective file source name. The
builder also increments the tests counters.

We also added JavaDoc for `IBundleSpectrum` and `IPackageSpectrum` that
was forgotten in the previous commit PR #55 (32cc818)