Skip Apache Maven GPG Plugin by default
Skipping the execution of `maven-gpg-plugin` by default. The plugin will
execute when the profile `release` is active. Using the property
`gpg.skip` to avoid write a lot of configurations.

The `release` profile is active on GitHub Actions CI to test if the
plugin will execute on the different operation systems and JDK versions.