Release v0.0.1 but using JDK 6
I tried to release with JDK 6 but the release process failed due to TLS
1.2 issues with Java 6. I was a bit on a rush and decided to release
using JDK 8 but obviously targeting Java 6 bytecode.

But all my work to make the build compatible with JDK 6 couldn't be for
nothing. I decided to tune my JDK 6 to allow it accepts TLS 1.2
connections.

Somethign similar to this Stack Overflow question:

https://stackoverflow.com/questions/33364100/how-to-use-tls-1-2-in-java-6/61115302#61115302

For some reason,  the default `cacerts` of JDK 6 provied by Apple was
not present on my machine, so I decided to use the same one provided by
Temurin 8.

It worked.

This sound too trick to me, so on next releases we will use at Java 8
and toolchains for JDK 6.