Module Jaguar2 Build
After pull requests #42, #43 and #44 the example modules were not
inherit Jaguar2 Parent POM, but this introduced a lot of duplicated
build configuration in the POM.

We tried to reduce the duplications and noisy configurations on pull
request #45, but since the Jaguar2 Examples do not inherit the Jaguar2
Parent POM, commom configuration are still duplicated.

For instance, to change the Surefire version on the entire project we
need to update two modules.

See commit a7fba58 and 155f1e3 from PR #69.

Also, we can forget to update a config on both places, like on commit
951a101 and commit d41855e from pull requests #63 and #68, respectively.

There we added support to JDK toolchain 6 and 7, but not updated the
Jaguar2 Examples module. So we are not building examples modules with
JDK 6/7.

That's why we are receiving the following warning message during the
build of the example modules with JDK 6, 7 and 8. Because we are only
using Java 8 in these cases.

```
[WARNING] bootstrap class path not set in conjunction with -source 1.6
```

See: https://github.com/saeg/jaguar2/issues/70

At least for JDK 6 this message should not happen.

This new module will extract common configurations from Jagar2 Parent
and Jaguar2 Examples modules.