Upgrade Apache Maven PMD Plugin 3.13.0 -> 3.21.2
This is the latest version of the plug-in.

This version targets Java 8 and rely on PMD version 6.55.0 that targets
Java 7.

See: https://github.com/apache/maven-pmd-plugin/blob/maven-pmd-plugin-3.21.2/pom.xml#L85-L86
See: https://maven.apache.org/plugins/maven-pmd-plugin/examples/upgrading-PMD-at-runtime.html#version-relationship-between-maven-pmd-plugin-and-pmd
See: https://github.com/pmd/pmd/blob/pmd_releases/6.55.0/pom.xml#L86

In other words: We need at least Java 8 to run the plug-in and at least
Java 7 to run PMD.

The plug-in supports Maven Toolchain since version 3.14.0.

See: PR #64 commit 4688afa

So, when running with JDK 6 toolchain (`-Djdk=6`) or with JDK 7
toolchain (`-Djdk=7`) PMD will execute with JDK 6 and 7, respectivelly.

This means that PMD will not work when running with JDK 6 toolchain as
it needs at least Java 7.

But, as you can see in the exception stack trace on 4688afa commit
message, looks that the PMD process is initialized by the plug-in class:
`org/apache/maven/plugins/pmd/exec/PmdExecutor`. And how the plug-in is
targetting Java 8, this means that the plug-in will not work when
running with JDK 6 and JDK 7 toolchains.

See: PR #76 for more details